fix import cycle problems by moving the enum into lib file
