+++ packages/shared/src/lib/referral.ts
@@ -1,0 +1,7 @@
+export enum ReferralCampaignKey {
+  Generic = 'generic',
+  Search = 'search',
+  SharePost = 'share_post',
+  ShareComment = 'share_comment',
+  ShareProfile = 'share_profile',
+}
